Update Source.cpp
Tarin here

I have moved my menu to a while loop that calls a parsing function. The function parses the input and executes other functions to print instructions, exit the program, or continue on to the rest of main. After that loop we pretend we are loading he user's file; as we are still in testing stage we just load up our sample file, but I have some placeholder lines in place for when we're loading up the user's actual file